@@ -6,9 +6,9 @@
   <name>SpaceRaze World</name>
   <description>SpaceRaze world obejcts</description>
   <properties>
-      <maven.compiler.target>13</maven.compiler.target>
-      <maven.compiler.source>13</maven.compiler.source>
-      <maven.complier.release>13</maven.complier.release>
+      <maven.compiler.target>21</maven.compiler.target>
+      <maven.compiler.source>21</maven.compiler.source>
+      <maven.complier.release>21</maven.complier.release>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
   </properties>
   <dependencies>
@@ -17,20 +17,20 @@
   		<artifactId>utils</artifactId>
   		<version>0.0.1-SNAPSHOT</version>
   	</dependency>
-  	<dependency>
-		<groupId>com.fasterxml.jackson.jaxrs</groupId>
-		<artifactId>jackson-jaxrs-json-provider</artifactId>
-		<version>2.9.5</version>
-	</dependency>
       <dependency>
-          <groupId>javax.persistence</groupId>
-          <artifactId>javax.persistence-api</artifactId>
-          <version>2.2</version>
+          <groupId>com.fasterxml.jackson.module</groupId>
+          <artifactId>jackson-module-jakarta-xmlbind-annotations</artifactId>
+          <version>2.17.0-rc1</version>
+      </dependency>
+      <dependency>
+          <groupId>jakarta.persistence</groupId>
+          <artifactId>jakarta.persistence-api</artifactId>
+          <version>3.2.0-M1</version>
       </dependency>
       <dependency>
           <groupId>org.projectlombok</groupId>
           <artifactId>lombok</artifactId>
-          <version>1.18.22</version>
+          <version>1.18.30</version>
       </dependency>
   </dependencies>
 </project>
