@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import spaceraze.world.enums.BlackMarketFrequency;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Setter
 @Getter
