@@ -6,7 +6,7 @@ import lombok.Getter;
 import lombok.Setter;
 import spaceraze.world.TurnInfo;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
