@@ -2,7 +2,7 @@ package spaceraze.world.mapinfo;
 
 import lombok.*;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.io.Serializable;
 
 /**
