@@ -8,7 +8,7 @@ import lombok.Setter;
 import spaceraze.world.Planet;
 import spaceraze.world.Spaceship;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Setter
 @Getter
