@@ -8,10 +8,7 @@ import lombok.Setter;
 import lombok.experimental.SuperBuilder;
 import spaceraze.world.enums.BattleGroupPosition;
 
-import javax.persistence.Entity;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
+import jakarta.persistence.*;
 
 @Setter
 @Getter
