@@ -6,9 +6,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.SuperBuilder;
-import spaceraze.world.Troop;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Setter
 @Getter
