@@ -2,16 +2,12 @@ package spaceraze.world.report.landbattle;
 
 import java.io.Serializable;
 
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.SuperBuilder;
 import spaceraze.world.enums.BattleGroupPosition;
-
-import javax.persistence.Entity;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
 
 @Setter
 @Getter
